pyproject.toml
[project]
name = "tg-bot-dict"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.1",
    "requests>=2.32.3",
]
